+++ packages/chatgpt/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-chatgpt
 
+## 1.0.1
+
+### Patch Changes
+
+- Expose documenatation, expand references and expose options.
+
 ## 1.0.0
 
 Initial release.