+++ packages/claude/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-claude
 
+## 1.0.2
+
+### Patch Changes
+
+- Expose documenatation, expand references and expose options.
+
 ## 1.0.1
 
 ### Patch Changes